+++ src/App.js
@@ -1,12 +1,14 @@
 import React from 'react';
 // import Main from './pages/main'; // Ensure the path is correct
-import Login from './pages/login'
+import Login from './pages/Login/login'
 import './App.css'
+import Header from './pages/Components/header'
 
 const App = () => {
     return (
         <div>
-            <Login /> {/* Use the Main component */}
+            <Login />
+            {/* <Header /> */}
         </div>
     );
 };
