+++ src/index.css
@@ -1,0 +1,82 @@
+.login_container {
+    font-family: 'Inter';
+    font-style: normal;
+    font-weight: 400;
+    background-color: #F0F0F0;
+    border-radius: 10px;
+    width: 450px;
+    height: 300px;
+    align-items: center;
+    justify-content: center;
+    display: flex;
+}
+
+.light-mode {
+    background-color: white;
+    color: #333; /* Darker gray for better readability */
+    transition: background-color 0.25s ease-in-out;
+}
+
+.dark-mode {
+    background-color: #181A1B;
+    color: white;
+    transition: background-color 0.25s ease-in-out;
+}
+
+.light-mode h1,
+.dark-mode h1,
+.light-mode p,
+.dark-mode p,
+.light-mode a,
+.dark-mode a,
+.light-mode label,
+.dark-mode label,
+.light-mode .brend-name-second-label,
+.dark-mode .brend-name-second-label,
+.light-mode .login-container,
+.dark-mode .login-container {
+    transition: color 0.25s ease-in-out;
+}
+
+.light-mode p, 
+.light-mode h1, 
+.light-mode a, 
+.light-mode label, 
+.light-mode button,
+.light_mode {
+    color: black; /* Use darker color for readability */
+}
+
+.light-mode .brend-name-first-label {
+    color: white;
+}
+
+.dark-mode input {
+    background-color: #181A1B;
+}
+
+.dark-mode p, 
+.dark-mode h1, 
+.dark-mode a, 
+.dark-mode label,
+.dark-mode input {
+    color: white; /* Text color in dark mode */
+}
+
+.dark-mode .brend-name-second-label {
+    color: black;
+}
+
+.dark-mode .login_container {
+    background-color: #1e2122;
+}
+
+.dark-mode .half-black-header, button {
+    color: white;
+    background-color: #181A1B;
+}
+
+.light-mode button {
+    color: white;
+    background-color: rgb(47, 110, 206);
+}
