+++ src/pages/Main-page/main.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 const Main = () => {
     return (
         <div>
-            asd
+            
         </div>
     );
 };
